+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 ### Changed
 
 - `Pendulum` class has been renamed to `DateTime`.
+- `strict` keyword argument for `parse` has been renamed to `exact`.
+- Made the `strict` keyword argument for `parse` false by default, which means it will not fallback on the `dateutil` parser.
 - Improved performances of the `precise_diff()` helper.